add error handling to grpc calling methods

+++ src/main/java/at/ac/fhcampuswien/se_booking/dto/currency/ConvertResultDTO.java
@@ -1,7 +1,7 @@
 package at.ac.fhcampuswien.se_booking.dto.currency;
 
 public record ConvertResultDTO(
-        Double amount,
+        Float amount,
         String currency
 ) {
 }
